refactoring: OtaPackageController: comments3
