+++ application/src/main/java/org/thingsboard/server/service/entitiy/otaPackageController/TbOtaPackageService.java
@@ -28,5 +28,5 @@ public interface TbOtaPackageService  {
     void  delete(OtaPackageInfo otaPackageInfo, SecurityUser user) throws ThingsboardException;
 
     OtaPackageInfo saveOtaPackageData(OtaPackageInfo otaPackageInfo, String checksum, ChecksumAlgorithm checksumAlgorithm,
-                                      byte[] data, String filename, String contentType, SecurityUser securityUser, Exception e) throws ThingsboardException;
+                                      byte[] data, String filename, String contentType, SecurityUser securityUser) throws ThingsboardException;
 }
